requirements.txt
# Core dependencies for web scraping framework
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Configuration management
PyYAML>=6.0
python-dotenv>=0.19.0

# Data export formats
python-docx>=0.8.11
openpyxl>=3.0.9

# Optional: For JavaScript-heavy sites (requires separate WebDriver setup)
# selenium>=4.0.0

# Development and testing (optional)
# pytest>=7.0.0
# pytest-cov>=4.0.0

# Logging and utilities
python-dateutil>=2.8.0